Moved Windows-only Permission function to Lidarr.Windows
